+++ sdk/keyvault/keyvault-keys/src/keysModels.ts
@@ -612,3 +612,85 @@ export interface GetCryptographyClientOptions {
    */
   keyVersion?: string;
 }
+
+/**
+ * Options for {@link KeyClient.rotateKey}
+ */
+export interface RotateKeyOptions extends coreHttp.OperationOptions {}
+
+/**
+ * The properties of a key rotation policy that the client can set for a given key.
+ *
+ * You may also reset the key rotation policy to its default values by setting lifetimeActions to an empty array.
+ */
+export interface KeyRotationPolicyProperties {
+  /**
+   * Optional key expiration period used to define the duration after which a newly rotated key will expire, defined as an ISO 8601 duration.
+   */
+  expiresIn?: string;
+
+  /**
+   * Actions that will be performed by Key Vault over the lifetime of a key.
+   *
+   * You may also pass an empty array to restore to its default values.
+   */
+  lifetimeActions?: KeyRotationLifetimeAction[];
+}
+
+/**
+ * The complete key rotation policy that belongs to a key.
+ */
+export interface KeyRotationPolicy extends KeyRotationPolicyProperties {
+  /**
+   * The identifier of the Key Rotation Policy.
+   * May be undefined if a policy has not been explicitly set.
+   */
+  readonly id?: string;
+
+  /**
+   * The created time in UTC.
+   * May be undefined if a policy has not been explicitly set.
+   */
+  readonly createdOn?: Date;
+
+  /**
+   * The last updated time in UTC.
+   * May be undefined if a policy has not been explicitly set.
+   */
+  readonly updatedOn?: Date;
+}
+
+/**
+ * An action and its corresponding trigger that will be performed by Key Vault over the lifetime of a key.
+ */
+export interface KeyRotationLifetimeAction {
+  /**
+   * Time after creation to attempt the specified action, defined as an ISO 8601 duration.
+   */
+  timeAfterCreate?: string;
+
+  /**
+   * Time before expiry to attempt the specified action, defined as an ISO 8601 duration.
+   */
+  timeBeforeExpiry?: string;
+
+  /**
+   * The action that will be executed.
+   */
+  action: KeyRotationPolicyAction;
+}
+
+/**
+ * The action that will be executed.
+ */
+export type KeyRotationPolicyAction = "Rotate" | "Notify";
+
+/**
+ * Options for {@link KeyClient.updateKeyRotationPolicy}
+ */
+export interface UpdateKeyRotationPolicyOptions extends coreHttp.OperationOptions {}
+
+/**
+ * Options for {@link KeyClient.getRotationPolicy}
+ */
+export interface GetKeyRotationPolicyOptions extends coreHttp.OperationOptions {}