+++ sdk/servicebus/service-bus/test/perf-js-libs/rhea-promise/send.ts
@@ -11,8 +11,11 @@ Measures the maximum throughput of `sender.send()` in package `rhea-promise`.
  */
 
 import { Connection, SenderEvents, ConnectionOptions } from "rhea-promise";
-import delay from "delay";
 import moment from "moment";
+
+function delay(ms: number): Promise<void> {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
 
 const _payload = Buffer.alloc(1024);
 const _start = moment();