+++ sdk/servicebus/service-bus/test/perf-js-libs/service-bus-v1/receive.ts
@@ -18,12 +18,15 @@ import {
   OnMessage,
   ServiceBusMessage
 } from "@azure/service-bus";
-import delay from "delay";
 import moment from "moment";
 
 // Load the .env file if it exists
 import * as dotenv from "dotenv";
 dotenv.config();
+
+function delay(ms: number): Promise<void> {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
 
 const _start = moment();